refactor: ImageContentType 패키지 이동
